Update README.md
Objective:

The primary goal of the HR Portal project is to create a secure and user-friendly web application that simplifies and provides a seamless execution of HR activities. This portal is designed to empower HR professionals to effectively oversee employee information, sanction leave applications, and simplify the employee onboarding process. Moreover, employees will be able to register, furnish their personal information, and access HR related services, all of which collectively enhance and streamline the HR processes within the organization.